업데이트
- [x]  보너스로 받는 재화는 floatingText에 + 해줘야함
- [x]  닫을때 소리
- [x]  업그레이드 텍스트 크기너무 큼
- [x]  바탕화면 눌렀을때 열려있는 패널 닫기
- [x]  몬스터 죽을때 소리가 안나는 이유는? Hit소리에 묻힘 → Hit Sound Pooling
- [x]  Damage float Spawn string → float로 변경

+++ Assets/WOM/Scripts/Component/Monster/DungeonMonster.cs
@@ -10,7 +10,7 @@ public class DungeonMonster : DungeonMonsterBase
     public DungenMonsterFileData curMonsterData;
 
     public int curLevel = 0;
-    public double curMonsterHP;
+    public float curMonsterHP;
     public string stageName;
     public DungeonMonsterData curData;
     public SpriteLibraryChanged spriteLibraryChanged;
